Update bankid auth api requests
- Refactor request building to separate file
- Add tests for building request
